changing js files to ts
