--- frontend/src/app/utils/message.js
@@ -1,7 +1,0 @@
-export const checkMessageLength = (msg, char) => {
-  if (msg.length > char) {
-    return msg.split('').slice(0, char).concat(['...']).join('');
-  } else {
-    return msg;
-  }
-};
