+++ frontend/src/app/utils/message.ts
@@ -1,0 +1,7 @@
+export const checkMessageLength = (msg: string, char: number) => {
+	if (msg.length > char) {
+		return msg.split('').slice(0, char).concat(['...']).join('');
+	} else {
+		return msg;
+	}
+};
