--- frontend/src/app/utils/scroll.js
@@ -1,4 +1,0 @@
-// scroll to the latest message
-export const scrollToBottom = (ref) => {
-  ref.current?.scrollIntoView({ behavior: 'smooth' });
-};
