+++ frontend/src/app/utils/scroll.ts
@@ -1,0 +1,4 @@
+// scroll to the latest message
+export const scrollToBottom = (ref: any) => {
+	ref.current?.scrollIntoView({ behavior: 'smooth' });
+};
