+++ frontend/src/app/utils/timestamp.ts
@@ -1,0 +1,7 @@
+import Sugar from 'sugar';
+
+// sugar date format
+export const createTimestamp = (formatString: string) => {
+	const formattedDate = Sugar.Date.format(new Date(), formatString);
+	return formattedDate;
+};
